Split the files along some boundaries
By splitting the single file into multiple files its easier to reason
about what is what.
